fix(eve): keep zod out of the workflow driver bundle
The driver graph acquired two runtime paths to the handle store's zod
schemas, inflating both function payloads by 1.4MB (zod plus inline
sourcemaps, base64-amplified into the embedded workflow code):

- turn-workflow -> handles/query -> store.safeParse. query.ts now reads
  the state key schema-free behind an Array.isArray guard; every write
  still validates through transitions.ts, and that trust boundary is
  documented on the reader.
- turn-workflow -> runtime-actions (for one pure helper) -> transitions
  -> store. resolveRuntimeActionResultsForKeys moves to
  runtime/actions/results.ts next to the key helpers it uses, so the
  driver no longer evaluates the write-side module at all.

AGENT_HANDLES_STATE_KEY moves to handles/state-key.ts (store re-exports
it) so the schema-free reader shares the canonical key without the
schema module. Weather fixture server bundle: 3,268,062 -> 1,745,826
bytes; the decoded workflow payload carries no zod.
