+++ packages/eve/src/harness/runtime-actions.ts
@@ -1,7 +1,8 @@
 import type { ModelMessage, ToolSet, TypedToolCall } from "ai";
 
 import { createActionResultEvent, type UnstampedMessageStreamEvent } from "#protocol/message.js";
-import { getRuntimeActionRequestKey, getRuntimeActionResultKey } from "#runtime/actions/keys.js";
+import { getRuntimeActionRequestKey } from "#runtime/actions/keys.js";
+import { resolveRuntimeActionResultsForKeys } from "#runtime/actions/results.js";
 import type { RuntimeActionRequest, RuntimeActionResult } from "#runtime/actions/types.js";
 import { parseJsonObject, type JsonObject } from "#shared/json.js";
 import type { TokenUsage } from "#shared/token-usage.js";
@@ -156,39 +157,6 @@ function resolveRuntimeActionResultsForBatch(input: {
     pendingKeys: input.batch.actions.map((action) => getRuntimeActionRequestKey(action)),
     results: input.results.filter((result) => isResultBoundToRunningHandle(input.state, result)),
   });
-}
-
-/** Returns results in pending-key order once every requested action has completed. */
-export function resolveRuntimeActionResultsForKeys(input: {
-  readonly pendingKeys: readonly string[];
-  readonly results: readonly RuntimeActionResult[];
-}): RuntimeActionResult[] | undefined {
-  const pendingKeySet = new Set(input.pendingKeys);
-  const resultsByKey = new Map<string, RuntimeActionResult>();
-
-  for (const result of input.results) {
-    const key = getRuntimeActionResultKey(result);
-
-    if (!pendingKeySet.has(key)) {
-      continue;
-    }
-
-    resultsByKey.set(key, result);
-  }
-
-  const orderedResults: RuntimeActionResult[] = [];
-
-  for (const key of input.pendingKeys) {
-    const result = resultsByKey.get(key);
-
-    if (result === undefined) {
-      return undefined;
-    }
-
-    orderedResults.push(result);
-  }
-
-  return orderedResults;
 }
 
 /**
